+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add `login_with` key in discopower definitions
+
+### Fixed
+
+- Load JavaScript scripts with the new way
+- Fix undefined index error for `idps_in_searchable_list_index`
+- Load jQuery CSS script
+
+### Changed
+
+- Move `ribbon_text` definition to config file
+
 ## [v3.0.0](https://github.com/rciam/simplesamlphp-module-themevanilla/compare/v2.3.1...v3.0.0) - 2019-09-20
 
 ### Added